move rejectIfSwapFail in utils.ts file
